+++ dev/resources/nginx.conf
@@ -66,6 +66,9 @@ http {
 
         # hmr
         proxy_read_timeout 86400;
+
+        # set client body size to 1 GB
+        client_max_body_size 1G;
         
         location = / {  
           return 302 /catalogs/dev;
